Layouting packet decoer; Adding communication handlers; Adding Text and VarInt values;

+++ octetmc-protocol/src/lib.rs
@@ -9,4 +9,4 @@ pub mod packet;
 pub const GAME_VERSIONS    : &[Version] = &[
     Version::new(1, 21, 5)
 ];
-pub const PROTOCOL_VERSION : usize = 770;
+pub const PROTOCOL_VERSION : u32 = 770;
